backend/requirements.txt
Python 3.11.7
Flask 2.2.5
Werkzeug 2.2.3
mysql.connector